fix: address remaining Copilot review comments on model discovery
- Snapshot dict items in ModelCache.get_all_cached() to prevent
  RuntimeError from concurrent mutation during iteration
- Move get_provider_kwargs to gateway.services.provider_kwargs to
  eliminate API-layer dependency from service layer (re-exported
  from chat.py for backward compatibility)
- Optimize get_model() to parse provider from model_id for targeted
  cache lookup instead of O(n) scan across all cached providers
