+++ src/gateway/services/provider_kwargs.py
@@ -1,0 +1,48 @@
+"""Shared helper for building provider kwargs from gateway configuration."""
+
+from typing import Any
+
+from any_llm import LLMProvider
+
+from gateway.auth.vertex_auth import setup_vertex_environment
+from gateway.core.config import GatewayConfig
+
+
+def get_provider_kwargs(
+    config: GatewayConfig,
+    provider: LLMProvider,
+) -> dict[str, Any]:
+    """Get provider kwargs from config for acompletion calls.
+
+    Args:
+        config: Gateway configuration
+        provider: Provider name
+
+    Returns:
+        Dictionary of provider kwargs (credentials, client_args, etc.)
+
+    """
+    kwargs: dict[str, Any] = {}
+    if provider.value in config.providers:
+        provider_config = config.providers[provider.value]
+
+        if provider == LLMProvider.VERTEXAI:
+            vertex_creds = provider_config.get("credentials")
+            vertex_project = provider_config.get("project")
+            vertex_location = provider_config.get("location")
+
+            kwargs.update(
+                setup_vertex_environment(
+                    credentials=vertex_creds,
+                    project=vertex_project,
+                    location=vertex_location,
+                )
+            )
+            if "client_args" in provider_config:
+                kwargs["client_args"] = provider_config["client_args"]
+        else:
+            kwargs = {k: v for k, v in provider_config.items() if k != "client_args"}
+            if "client_args" in provider_config:
+                kwargs["client_args"] = provider_config["client_args"]
+
+    return kwargs
